Moving GetReviewerPapersAsync to the Reviewer Assignment Service
