+++ Reviewer2/Reviewer2.Services/ReviewAssignments/IReviewAssignmentService.cs
@@ -101,4 +101,12 @@ public interface IReviewAssignmentService
     /// The total number of review assignments created.
     /// </returns>
     Task<int> AutoAssignReviewersAsync(int reviewersPerPaper);
+    
+    /// <summary>
+    /// Retrieves all papers assigned to a reviewer, including review status and details.
+    /// </summary>
+    /// <param name="reviewerId">The unique identifier of the reviewer.</param>
+    /// <returns>A list of <see cref="ReviewerPaperDTO"/> representing all review assignments.</returns>
+    /// <exception cref="ArgumentException">Thrown if <paramref name="reviewerId"/> is empty.</exception>
+    Task<List<ReviewerPaperDTO>> GetReviewerPapersAsync(Guid reviewerId);
 }
