+++ Reviewer2/Reviewer2.Services/Submissions/PaperSubmission/IPaperQueryService.cs
@@ -75,13 +75,5 @@ namespace Reviewer2.Services.Submissions.PaperSubmission
         /// and file summaries suitable for display in an overview table.
         /// </returns>
         Task<IEnumerable<UserPaperDTO>> GetAllPapersAsync(PaperStatus? status = null);
-
-        /// <summary>
-        /// Retrieves all papers assigned to a reviewer, including review status and details.
-        /// </summary>
-        /// <param name="reviewerId">The unique identifier of the reviewer.</param>
-        /// <returns>A list of <see cref="ReviewerPaperDTO"/> representing all review assignments.</returns>
-        /// <exception cref="ArgumentException">Thrown if <paramref name="reviewerId"/> is empty.</exception>
-        Task<List<ReviewerPaperDTO>> GetReviewerPapersAsync(Guid reviewerId);
     }
 }
