+++ Reviewer2/Reviewer2.Services/Submissions/PaperSubmission/PaperQueryService.cs
@@ -308,37 +308,4 @@ public class PaperQueryService : IPaperQueryService
 
         return result;
     }
-    
-    /// <inheritdoc/>
-    public async Task<List<ReviewerPaperDTO>> GetReviewerPapersAsync(Guid reviewerId)
-    {
-        if (reviewerId == Guid.Empty)
-            throw new ArgumentException("Reviewer ID cannot be empty.", nameof(reviewerId));
-
-        Log.Information("Fetching all review assignments for reviewer {ReviewerId}", reviewerId);
-
-        await using var dbContext = await _contextFactory.CreateDbContextAsync();
-
-        // Eagerly load related data for mapping
-        var assignments = await dbContext.ReviewAssignments
-            .Where(ra => ra.ReviewerId == reviewerId)
-            .Include(ra => ra.Review)
-            .Include(ra => ra.Paper)
-            .ThenInclude(p => p.Authors)
-            .ThenInclude(a => a.User)
-            .Include(ra => ra.Paper)
-            .ThenInclude(p => p.Files)
-            .ToListAsync();
-
-        Log.Information("Retrieved {Count} review assignments for reviewer {ReviewerId}", assignments.Count, reviewerId);
-
-        // Map assignments to DTOs using the extension method
-        var dtos = assignments
-            .Select(ra => ra.ToReviewerPaperDto())
-            .ToList();
-
-        Log.Information("Mapped {Count} review assignments to ReviewerPaperDTO for reviewer {ReviewerId}", dtos.Count, reviewerId);
-
-        return dtos;
-    }
 }
